+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/translate/components/TranslateWithGroups.tsx
@@ -140,6 +140,7 @@ const Element = ({
     FormElementTypes.formattedDate,
     FormElementTypes.addressComplete,
     FormElementTypes.fileInput,
+    FormElementTypes.starRating,
   ];
 
   if (element.type === FormElementTypes.dynamicRow) {