+++ lib/responseDownloadFormats/html/components/RowTable.tsx
@@ -7,6 +7,7 @@ import { FormElementTypes, FormRecord } from "@lib/types";
 import { formatUserInput } from "@lib/utils/strings";
 import { Language } from "@lib/types/form-builder-types";
 import { formatNumberInputAnswer } from "@lib/responseDownloadFormats/utils/formatNumberInputAnswer";
+import { formatStarRatingAnswer } from "@lib/responseDownloadFormats/utils/formatStarRatingAnswer";
 
 const QuestionRows = ({
   submission,
@@ -20,6 +21,7 @@ const QuestionRows = ({
   const { t } = customTranslate("common");
   const renderColumn = (index: number, lang: Language, item: Answer, subItem = false) => {
     const numberInputValue = formatNumberInputAnswer(item, lang, formRecord);
+    const starRatingValue = formatStarRatingAnswer(item, formRecord);
     return (
       <div
         key={`row-${index}`}
@@ -43,6 +45,8 @@ const QuestionRows = ({
         <dd className="p-4">
           {item.type === FormElementTypes.numberInput ? (
             <p>{numberInputValue}</p>
+          ) : starRatingValue !== undefined ? (
+            <p>{starRatingValue}</p>
           ) : (
             <p dangerouslySetInnerHTML={{ __html: formatUserInput(String(item.answer)) }}></p>
           )}