+++ packages/types/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.38] - 2026-06-23
+
+- Add Star Rating element type
+
 ## [1.0.37] - 2026-07-22
 
 - Republish package following failed NPM publication
@@ -52,7 +56,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.25] - 2025-10-17
 
 - Remove File Upload as a beta option
-
 
 ## [1.0.23] - 2025-10-01